Finished view event and attending events
view event is essentially done, but you have to manually load the event
page. It will show the hidden features and just load the data from the
previous pages.

Also added templates, there mainly for features for that one account, if
you need to differentiate between the three your gonna have to copy the
account.php.